% Auto-generated: do not edit by hand
\name{barChart}

\alias{barChart}

\title{BarChart component}

\description{
BarChart is an chartXkcd-react component. It takes `labels`, and `dataset` as inputs. displays a BarChart. It also takes title,xLabel,yLabel,options as optional inputs. user can pass chart.xkcd options except the legend positions
}

\usage{
barChart(id=NULL, dataset=NULL, labels=NULL, options=NULL,
title=NULL, xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataset}{Unnamed list. The dataset for BarChart.}

\item{labels}{Unnamed list. A labels for BarChart.}

\item{options}{Unnamed list. Options for the BarChart.}

\item{title}{Character. The Title for BarChart.}

\item{xLabel}{Character. The XLabel for BarChart.}

\item{yLabel}{Character. The YLabel for BarChart.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

