% Auto-generated: do not edit by hand
\name{lineChart}

\alias{lineChart}

\title{LineChart component}

\description{
LineChart is an chartXkcd-react component. It takes `labels`, and `dataset` as inputs. displays a LineChart. It also takes title,xLabel,yLabel,options as optional inputs. user can pass chart.xkcd options except the legend positions
}

\usage{
lineChart(id=NULL, dataset=NULL, labels=NULL, options=NULL,
title=NULL, xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataset}{Unnamed list. The dataset for LineChart.}

\item{labels}{Unnamed list. A labels for lineChart.}

\item{options}{Unnamed list. Options for the LineChart.}

\item{title}{Character. The Title for LineChart.}

\item{xLabel}{Character. The XLabel for LineChart.}

\item{yLabel}{Character. The YLabel for LineChart.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

