% Auto-generated: do not edit by hand
\name{pieChart}

\alias{pieChart}

\title{PieChart component}

\description{
PieChart is an chartXkcd-react component. It takes `labels`, and `dataset` as inputs. displays a PieChart. It also takes title,xLabel,yLabel,options as optional inputs. user can pass chart.xkcd options except the legend positions
}

\usage{
pieChart(id=NULL, dataset=NULL, labels=NULL, options=NULL,
title=NULL, xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataset}{Unnamed list. The dataset for PieChart.}

\item{labels}{Unnamed list. A labels for PieChart.}

\item{options}{Unnamed list. Options for the PieChart.}

\item{title}{Character. The Title for PieChart.}

\item{xLabel}{Character. The XLabel for PieChart.}

\item{yLabel}{Character. The YLabel for PieChart.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

