% Auto-generated: do not edit by hand
\name{radarChart}

\alias{radarChart}

\title{RadarChart component}

\description{
RadarChart is an chartXkcd-react component. It takes `labels`, and `dataset` as inputs. displays a RadarChart. It also takes title,xLabel,yLabel,options as optional inputs. user can pass chart.xkcd options except the legend positions
}

\usage{
radarChart(id=NULL, dataset=NULL, labels=NULL, options=NULL,
title=NULL, xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataset}{Unnamed list. The dataset for RadarChart.}

\item{labels}{Unnamed list. A labels for RadarChart.}

\item{options}{Unnamed list. Options for the RadarChart.}

\item{title}{Character. The Title for RadarChart.}

\item{xLabel}{Character. The XLabel for RadarChart.}

\item{yLabel}{Character. The YLabel for RadarChart.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

