% Auto-generated: do not edit by hand
\name{scatterChart}

\alias{scatterChart}

\title{ScatterChart component}

\description{
ScatterChart is an chartXkcd-react component. It takes `labels`, and `dataset` as inputs. displays a ScatterChart. It also takes title,xLabel,yLabel,options as optional inputs. user can pass chart.xkcd options except the legend positions
}

\usage{
scatterChart(id=NULL, dataset=NULL, options=NULL, title=NULL,
xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataset}{Unnamed list. The dataset for ScatterChart.}

\item{options}{Unnamed list. Options for the ScatterChart.}

\item{title}{Character. The Title for ScatterChart.}

\item{xLabel}{Character. The XLabel for ScatterChart.}

\item{yLabel}{Character. The YLabel for ScatterChart.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

