% Auto-generated: do not edit by hand
\name{stackedChart}

\alias{stackedChart}

\title{StackedChart component}

\description{
StackedBarChart is an chartXkcd-react component. It takes `labels`, and `dataset` as inputs. displays a StackedBarChart. It also takes title,xLabel,yLabel,options as optional inputs. user can pass chart.xkcd options except the legend positions
}

\usage{
stackedChart(id=NULL, dataset=NULL, labels=NULL, options=NULL,
title=NULL, xLabel=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataset}{Unnamed list. The dataset for StackedBarChart.}

\item{labels}{Unnamed list. A labels for StackedBarChart.}

\item{options}{Unnamed list. Options for the StackedBarChart.}

\item{title}{Character. The Title for StackedBarChart.}

\item{xLabel}{Character. The XLabel for StackedBarChart.}

\item{yLabel}{Character. The YLabel for StackedBarChart.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

